+++ src/store/CHANGELOG.md
@@ -7,6 +7,8 @@ CHANGELOG
  * Add `StoreInterface::remove()` method
  * [BC BREAK] Make `Indexer` stateless - removed `$source` constructor parameter and `withSource()` method
  * Add `ConfiguredIndexer` decorator for pre-configuring default sources
+ * [BC BREAK] Reorder `Indexer` constructor parameters and make `$loader` optional
+ * Add support for passing documents directly to `Indexer::index()` without requiring a loader
 
 0.3
 ---